Added a memoizing abstraction to the jQuery.getJSON function.
The abstraction function getCachedJSON caches the fetched data
to window.localStorage and always checks the cache first to avoid
unnecessary network connections.

TODO: Add some logic to clear to cache when it is full (might be
hard since the browsers don't (?) send any event to notify that
space is about to run out or has run out). Also add logic to
try a few times if the fetching times out so that brief network
interruptions don't produce an error right away.